Finish ruling out potential IMLS plots; random order of candidates
Adjusted buffer calculations so that the buffer goes around the
intended square plot, which means needing an approximate 14 m buffer
for the plot (plus the appropriate buffer for the feature); then rule
out anything that was harvested, stratify by burns, and remove and
plots by group that aren’t clustered with the others.  Next step is to
go out and what plots look odd (highly invaded, etc).