Trying to create date difference
Trying to create a way to calculatre the rate of leaf litter per day but am now stuck on the step of creating the values of the difference between days. Comparing them while in the same column has seemed more complicated so I've tried creating a new column but that is currently also not working.