Added slope calculations and fixed date_comp
When trying to create a loop to determine the slope of some values I found out my date_comp was previously always assuming that a species had a sample at the most recent date. This was throwing off some numbers so I fixed it and added the slope section. Also some general cleaning and graphically additions.